+++ migrations/versions/f7ac3d27bb1d_notifications.py
@@ -21,7 +21,7 @@ def upgrade():
     op.create_table('notification',
     sa.Column('id', sa.Integer(), nullable=False),
     sa.Column('name', sa.String(length=128), nullable=True),
-    sa.Column('user_id', sa.Integer(), nullable=True),
+    sa.Column('user_id', sa.BigInteger(), nullable=True),
     sa.Column('timestamp', sa.Float(), nullable=True),
     sa.Column('payload_json', sa.Text(), nullable=True),
     sa.ForeignKeyConstraint(['user_id'], ['user.id'], ),
